+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
 resolver = "2"
-members = ["ddk", "bella/src-tauri", "payouts", "ddk-sample"]
+members = ["ddk", "bella/src-tauri", "payouts"]
