+++ ddk/Cargo.toml
@@ -45,6 +45,6 @@ sled = "0.34.7"
 async-trait = "0.1.80"
 
 [dev-dependencies]
-electrsd = { version = "0.22.0", features = ["legacy", "esplora_a33e97e1", "bitcoind_23_0"] }
+# electrsd = { version = "0.22.0", features = ["legacy", "esplora_a33e97e1", "bitcoind_23_0"] }
 electrum-client = "0.12.0"
 futures = "0.3.29"
